+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 * Your contribution here
+- [#23](https://github.com/numbata/grape-oas/pull/23): Add oneOf support for response schemas [@slbug](https://github.com/slbug)
 - [#22](https://github.com/numbata/grape-oas/pull/22): Handle boolean types in dry introspector [@slbug](https://github.com/slbug)
 
 ## [1.0.3] - 2025-12-23